refactor: Bearer 토큰 추출 중복 로직을 별도의 클래스에서 관리하도록 수정
